separate out triton inference logic from inference stage logic
