+++ tests/examples/log_parsing/test_inference.py
@@ -168,24 +168,21 @@ def test_log_parsing_triton_inference_log_parsing_build_output_message(config: C
 def test_log_parsing_inference_stage_constructor(config: Config, import_mod: typing.List[types.ModuleType]):
     inference_mod = import_mod[0]
 
-
     expected_inout_mapping = {}
-    expected_inout_mapping.update(inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP, {}).get("inputs", {}))
-    expected_inout_mapping.update(inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP, {}).get("outputs", {}))
+    expected_inout_mapping.update(
+        inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP,
+                                                                                           {}).get("inputs", {}))
+    expected_inout_mapping.update(
+        inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP,
+                                                                                           {}).get("outputs", {}))
 
     expected_kwargs = {
-        "model_name":
-            'test_model',
-        "server_url":
-            'test_server',
-        "force_convert_inputs":
-            False,
-        "use_shared_memory":
-            False,
-        "needs_logits":
-            True,
-        "inout_mapping":
-            expected_inout_mapping,
+        "model_name": 'test_model',
+        "server_url": 'test_server',
+        "force_convert_inputs": False,
+        "use_shared_memory": False,
+        "needs_logits": True,
+        "inout_mapping": expected_inout_mapping,
     }
 
     stage = inference_mod.LogParsingInferenceStage(
@@ -212,8 +209,12 @@ def test_log_parsing_inference_stage_get_inference_worker(config: Config, import
                                                    inout_mapping={'test': 'this'})
 
     expected_inout_mapping = {}
-    expected_inout_mapping.update(inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP, {}).get("inputs", {}))
-    expected_inout_mapping.update(inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP, {}).get("outputs", {}))
+    expected_inout_mapping.update(
+        inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP,
+                                                                                           {}).get("inputs", {}))
+    expected_inout_mapping.update(
+        inference_mod.LogParsingInferenceStage._INFERENCE_WORKER_DEFAULT_INOUT_MAPPING.get(PipelineModes.NLP,
+                                                                                           {}).get("outputs", {}))
     expected_inout_mapping.update({'test': 'this'})
 
     worker = stage._get_inference_worker(inf_queue=ProducerConsumerQueue())
