+++ tests/test_abp.py
@@ -139,9 +139,7 @@ def test_abp_cpp(config, tmp_path):
     # We are feeding TritonInferenceStage the port to the grpc server because that is what the validation tests do
     # but the code under-the-hood replaces this with the port number of the http server
     pipe.add_stage(
-        TritonInferenceStage(config,
-                             model_name='abp-nvsmi-xgb',
-                             server_url='localhost:8001',
+        TritonInferenceStage(config, model_name='abp-nvsmi-xgb', server_url='localhost:8001',
                              force_convert_inputs=True))
     pipe.add_stage(MonitorStage(config, description="Inference Rate", smoothing=0.001, unit="inf"))
     pipe.add_stage(AddClassificationsStage(config))
